refactor: change component from dash notation to camel
