+++ pahlava/components/Navbar.tsx
@@ -1,25 +1,19 @@
-import styled from "@emotion/styled";
 import {
   AppBar,
   Avatar,
-  Badge,
   Box,
   IconButton,
   Toolbar,
-  Tooltip,
   useTheme,
 } from "@mui/material";
 import MenuIcon from "@mui/icons-material/Menu";
-import SearchIcon from "@mui/icons-material/Search";
-import { Bell as BellIcon } from "../icons/bell";
 import { UserCircle as UserCircleIcon } from "../icons/user-circle";
-import { Users as UsersIcon } from "../icons/users";
 
 interface Props {
   onSidebarOpen: () => void;
 }
 
-export const DashboardNavbar = (props: Props) => {
+export const Navbar = (props: Props) => {
   const { onSidebarOpen, ...other } = props;
   const theme = useTheme();
   return (
