+++ pahlava/components/ProjectList.tsx
@@ -1,0 +1,7 @@
+const ProjectList = () => {
+  return (
+    <div>ProjectList</div>
+  )
+}
+
+export default ProjectList
